requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
websockets==14.1
pydantic==2.10.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
watchfiles==1.0.0
jinja2==3.1.4
aiohttp==3.11.11
pyyaml==6.0.2
httpx==0.28.1
